srcs/containers/transcendence/requirements.txt
# requirements.txt

djangorestframework==3.15.2

# Django-related packages
Django==5.0.6                   # The Django web framework
django-formtools==2.5.1         # Collection of assorted utilities for forms
django-otp==1.5.2               # One-time password framework for Django
django-phonenumber-field==8.0.0 # Phone number field for Django
phonenumbers
django-request==1.5.6           # Django middleware for request logging
django-two-factor-auth==1.17.0  # Two-factor authentication for Django
sqlparse==0.5.1                 # SQL parser used by Django's ORM

# Django Channels and related packages
django-redis
channels==4.1.0                 # Asynchronous support for Django
channels-redis==4.2.0           # Redis-backed channel layers
daphne==4.1.2                   # HTTP, HTTP2, and WebSocket protocol server for ASGI and ASGI-HTTP
uvicorn==0.31.0                 
h11==0.14.0
click==8.1.7

# WebSocket and asynchronous support
asgiref==3.8.1                  # ASGI (Asynchronous Server Gateway Interface) reference implementation
autobahn==23.6.2                # WebSocket and WAMP library
txaio==23.1.1                   # Compatibility layer for Twisted/Trollius for Autobahn
async-timeout==4.0.3            # Timeout context manager for asyncio
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
httptools==0.6.1
exceptiongroup==1.2.2
anyio==4.6.0
django-widget-tweaks==1.5.0

# Twisted-related packages
Twisted==24.3.0                 # Event-driven networking engine
Automat==22.10.0                # Self-service finite-state machines for Twisted
constantly==23.10.4             # Symbolic constants library
incremental==22.10.0            # Library for versioning Python projects
service-identity==24.1.0        # Service identity verification for Twisted
zope.interface==6.4.post2       # Interfaces for Python, used by Twisted

# Security and cryptography
cryptography==43.0.0            # Cryptographic recipes and primitives
pyOpenSSL==24.2.1               # Python wrapper for OpenSSL
pyasn1==0.6.0                   # ASN.1 library for Python
pyasn1_modules==0.4.0           # A collection of ASN.1 data structures
service-identity==24.1.0        # Service identity verification (already listed under Twisted)
pycryptodome==3.20.0

# HTTP and requests
requests==2.28.2                # HTTP library for Python
idna==3.7                       # Internationalized Domain Names in Applications (used by requests)
hyperlink==21.0.0               # Immutable, extensible URL class (used by Twisted and requests)

# Redis
redis==5.0.7                    # Redis-py, a Python client for Redis
redlock-py

# PostgreSQL
psycopg2-binary==2.9.9          # PostgreSQL adapter for Python
psycopg2-pool==1.2              # Connection pool for psycopg2

# Miscellaneous
attrs==23.2.0                   # Classes without boilerplate
cffi==1.16.0                    # Foreign Function Interface for Python calling C code
pycparser==2.22                 # C parser in Python (used by cffi)
six==1.16.0                     # Python 2 and 3 compatibility library
typing_extensions==4.12.2       # Backports and extensions of the typing module
msgpack==1.0.8                  # MessagePack (efficient binary serialization) for Python

# Image processing and QR codes
Pillow                          # Python Imaging Library (PIL) fork
pypng==0.20220715.0             # Pure Python PNG image encoder/decoder
qrcode==7.4.2                   # QR Code generator

